Completely detach irc->matrix processing from bridge.js
Now lives in irc-to-matrix.js. Each direction (i->m and m->i)
are distinct from each other, so having them in the same file
made no sense. This will allow easier code reloading for
specific directions on the bridge.
