+++ lib/irc-appservice.js
@@ -8,6 +8,7 @@ module.exports.serviceName = "matrix-appservice-irc";
 var q = require("q");
 
 var bridge = require("./bridge.js");
+var ircToMatrix = require("./irc-to-matrix.js");
 var IrcServer = require("./irclib/server.js").IrcServer;
 var ircLib = require("./irclib/irc.js");
 var matrixLib = require("./mxlib/matrix");
@@ -187,10 +188,10 @@ module.exports.register = function(controller, serviceConfig) {
         localpart: asLocalpart
     });
     ircLib.registerHooks({
-        onMessage: bridge.hooks.irc.onMessage,
-        onJoin: bridge.hooks.irc.onJoin,
-        onPart: bridge.hooks.irc.onPart,
-        onMode: bridge.hooks.irc.onMode
+        onMessage: ircToMatrix.onMessage,
+        onJoin: ircToMatrix.onJoin,
+        onPart: ircToMatrix.onPart,
+        onMode: ircToMatrix.onMode
     });
     if (!serviceConfig.generateRegistration) {
         log.info("Joining mapped Matrix rooms...");
